Fixed bugs in NumStepper and Sequencer
Fixed issue in NumStepper where when typing a new tempo (like 120), as
soon as the 1 is entered the bpm is changed to 1 and it takes forever to
get to the next beat. I made it so you can't directly type in a new bpm
value, you have to use the spinner buttons.

I also fixed a bug with the audio context. When starting the sequence
without having pressed any other buttons, the draw loop would get stuck
on the first beat since the audio context was in a suspended state. This
was because i was checking if this.audioContext === "suspended" instead
of this.audioContext.state === "suspended"